RELEASE 1.5.0 update package librespot-java to 1.5.1

+++ src/main/java/me/haga/librespot/spotifi/model/CurrentSong.java
@@ -10,7 +10,7 @@ public class CurrentSong {
 
 
     private String current;
-    private Long trackTime;
+    private int trackTime;
     @JsonSerialize(using = ProtobufSerializer.class)
     private Metadata.Episode episode;
     @JsonSerialize(using = ProtobufSerializer.class)
@@ -43,11 +43,11 @@ public class CurrentSong {
         this.current = current;
     }
 
-    public Long getTrackTime() {
+    public int getTrackTime() {
         return trackTime;
     }
 
-    public void setTrackTime(Long trackTime) {
+    public void setTrackTime(int trackTime) {
         this.trackTime = trackTime;
     }
